requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.2.10
asn1crypto==0.24.0
astroid==2.4.2
Brlapi==0.6.6
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0    
dataclasses==0.7
defer==1.0.6
distlib==0.3.0
distro-info===0.18ubuntu0.18.04.1
Django==3.0.7
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==2.3.0
django-six==1.0.4
djangorestframework==3.11.0
falcon==3.0.0a1
fastapi==0.58.0
filelock==3.0.12
flake8==3.8.3
flake8-docstrings==1.5.0
gunicorn==20.0.4
h11==0.9.0
httpie==2.1.0
httplib2==0.9.2
httptools==0.1.1
idna==2.9
importlib-metadata==1.7.0
importlib-resources==2.0.1
isort==4.3.21
Jinja2==2.11.2
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
multitasking==0.0.9
netifaces==0.10.4
numpy==1.18.5
oauth==1.0.1
olefile==0.45.1
packaging==20.4
pandas==1.0.4
pexpect==4.2.1
Pillow==5.1.0
pluggy==0.13.1
protobuf==3.0.0
psycopg2-binary==2.8.3
py==1.8.1
pycairo==1.16.2
pycodestyle==2.6.0
pycrypto==2.6.1
pycups==1.9.73
pydantic==1.5.1
pydocstyle==5.0.2
pyflakes==2.2.0
Pygments==2.6.1
pygobject==3.26.1
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.23.0
requests-unixsocket==0.1.5
rope==0.17.0
screen==1.0.1
SecretStorage==2.3.1
simplejson==3.13.2
six==1.15.0
snowballstemmer==2.0.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
starlette==0.13.4
system-service==0.3
systemd-python==234
toml==0.10.1
tox==3.15.2
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.3
uvicorn==0.11.5
uvloop==0.14.0
uWSGI==2.0.18
virtualenv==20.0.23
wadllib==1.3.2
waitress==1.4.4
websockets==8.1
wrapt==1.12.1
xkit==0.0.0
xlwt==1.3.0
yfinance==0.1.54
zipp==3.1.0
zope.interface==4.3.2
